split Backend.js into classes
